+++ src/main/java/com/petros/bookstore/service/BookService.java
@@ -108,7 +108,7 @@ public class BookService {
      * @param pageable pagination information
      * @return a page of book responses matching the filters
      */
-    public Page<BookResponse> searchBooks(String title, String author, Integer availability, Genre genre, Float minPrice, Float maxPrice, Pageable pageable) {
+    public Page<BookResponse> searchBooks(String title, String author, Integer availability, Genre genre, Double minPrice, Double maxPrice, Pageable pageable) {
         return bookRepository.searchBooks(title, author, genre, availability, minPrice, maxPrice, pageable)
                 .map(BookMapper::toResponse);
     }
